Half working memory store replaced by redis
