+++ scalastyle-config.xml
@@ -46,11 +46,11 @@
     <check level="error" class="org.scalastyle.scalariform.NoCloneChecker" enabled="true"></check>
     <check level="error" class="org.scalastyle.scalariform.NoFinalizeChecker" enabled="true"></check>
     <check level="error" class="org.scalastyle.scalariform.CovariantEqualsChecker" enabled="true"></check>
-    <check level="error" class="org.scalastyle.file.RegexChecker" enabled="true">
-        <parameters>
-            <parameter name="regex"><![CDATA[println]]></parameter>
-        </parameters>
-    </check>
+    <!--<check level="error" class="org.scalastyle.file.RegexChecker" enabled="true">-->
+        <!--<parameters>-->
+            <!--<parameter name="regex"><![CDATA[println]]></parameter>-->
+        <!--</parameters>-->
+    <!--</check>-->
     <check level="error" class="org.scalastyle.scalariform.NumberOfTypesChecker" enabled="true">
         <parameters>
             <parameter name="maxTypes"><![CDATA[70]]></parameter>
